# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/orders:
    get:
      tags: [admin]
      operationId: adminListOrders
      summary: List all orders (admin)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Order" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /admin/orders/{id}:
    get:
      tags: [admin]
      operationId: adminGetOrder
      summary: Get an order
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    patch:
      tags: [admin]
      operationId: adminUpdateOrderStatus
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderStatusUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }

  # ── Admin: Overview ───────────────────────────────────────────────────
...
  /admin/homepage-seo:
    get:
      tags: [admin]
      operationId: adminGetHomepageSeo
      summary: Read the homepage SEO singleton (raw + effective with defaults)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HomepageSeoEnvelope" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    patch:
      tags: [admin]
      operationId: adminPatchHomepageSeo
      summary: Update the homepage SEO singleton (partial)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HomepageSeoPatch" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HomepageSeoEnvelope" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /admin/homepage-seo/suggest:
    post:
      tags: [admin]
      operationId: adminSuggestHomepageSeo
      summary: AI-suggest a fresh H1, sub-line, sub-copy and meta keywords
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HomepageSeoSuggestion" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    ShippingAddress:
      type: object
      required: [line1, city, postalCode, country]
      properties:
        line1: { type: string }
        line2: { type: string }
        city: { type: string }
        region: { type: string }
        postalCode: { type: string }
        country: { type: string }
    OrderItem:
      type: object
      required: [productId, productName, variantSize, quantity, unitPrice]
      properties:
        productId: { type: string }
        productName: { type: string }
        variantSize: { type: string }
        quantity: { type: integer }
        unitPrice: { type: integer }
        subscription: { type: boolean }
    Order:
      type: object
      required: [id, customerName, customerEmail, shippingAddress, currency, subtotal, total, status, items, createdAt]
      properties:
        id: { type: integer }
        customerName: { type: string }
        customerEmail: { type: string }
        shippingAddress: { $ref: "#/components/schemas/ShippingAddress" }
        currency: { type: string }
        subtotal: { type: integer }
        total: { type: integer }
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]
        notes:
          type: string
          nullable: true
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        createdAt: { type: string, format: date-time }
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]
        notes:
          type: string
          nullable: true
...
    HomepageSeoEffective:
      type: object
      required:
        - h1Visible
        - h1SrOnly
        - h1Subline
        - heroEyebrow
        - heroSubcopy
        - metaKeywords
        - breadcrumbHomeLabel
        - breadcrumbsJsonLdEnabled
      properties:
        h1Visible: { type: string }
        h1SrOnly: { type: string }
        h1Subline: { type: string }
        heroEyebrow: { type: string }
        heroSubcopy: { type: string }
        metaKeywords:
          type: array
          items: { type: string }
        breadcrumbHomeLabel: { type: string }
        breadcrumbsJsonLdEnabled: { type: boolean }
    HomepageSeoRaw:
      type: object
      required: [id, breadcrumbsJsonLdEnabled, metaKeywords, updatedAt]
      properties:
        id: { type: integer }
        h1Visible: { type: string, nullable: true }
        h1SrOnly: { type: string, nullable: true }
        h1Subline: { type: string, nullable: true }
        heroEyebrow: { type: string, nullable: true }
        heroSubcopy: { type: string, nullable: true }
        metaKeywords:
          type: array
          items: { type: string }
        breadcrumbHomeLabel: { type: string, nullable: true }
        breadcrumbsJsonLdEnabled: { type: boolean }
        lastAiSuggestionAt: { type: string, format: date-time, nullable: true }
        aiNotes: { type: string, nullable: true }
        updatedAt: { type: string, format: date-time }
    HomepageSeoEnvelope:
      type: object
      required: [raw, effective]
      properties:
        raw: { $ref: "#/components/schemas/HomepageSeoRaw" }
        effective: { $ref: "#/components/schemas/HomepageSeoEffective" }
    HomepageSeoPatch:
      type: object
      properties:
        h1Visible: { type: string, nullable: true, maxLength: 200 }
        h1SrOnly: { type: string, nullable: true, maxLength: 800 }
        h1Subline: { type: string, nullable: true, maxLength: 240 }
        heroEyebrow: { type: string, nullable: true, maxLength: 80 }
        heroSubcopy: { type: string, nullable: true, maxLength: 400 }
        metaKeywords:
          type: array
          maxItems: 20
          items: { type: string, minLength: 2, maxLength: 80 }
        breadcrumbHomeLabel: { type: string, nullable: true, maxLength: 40 }
        breadcrumbsJsonLdEnabled: { type: boolean }
        aiNotes: { type: string, nullable: true, maxLength: 800 }
    HomepageSeoSuggestion:
      type: object
      required: [suggestion, aiAvailable]
      properties:
        aiAvailable: { type: boolean }
        suggestion:
          type: object
          required:
            - h1Visible
            - h1SrOnly
            - h1Subline
            - heroEyebrow
            - heroSubcopy
            - metaKeywords
            - rationale
          properties:
            h1Visible: { type: string }
            h1SrOnly: { type: string }
            h1Subline: { type: string }
            heroEyebrow: { type: string }
            heroSubcopy: { type: string }
            metaKeywords:
              type: array
              items: { type: string }
            rationale: { type: string }